git補完

+++ install.sh
@@ -6,7 +6,10 @@ do
     [[ "$F" == ".gitignore" ]] && continue
     [[ "$F" == ".DS_Store" ]] && continue
 
-    echo "$F"
+    [[ -e ~/$F ]] && continue
+    echo "create $F"
     ln -s `pwd`/$F ~/$F
-    exit
 done
+
+[[ ! -e ~/local ]] && mkdir ~/local
+[[ ! -e ~/pkg ]] && mkdir ~/pkg
